requirements.txt
pandas
numpy
scikit-learn
matplotlib.pyplot
seaborn
